refactoring: BettingStatsDisplay 공용 컴포넌트로 이동 [#60]
