+++ frontend/src/shared/components/BettingStatsDisplay/BettingStatsDisplay.tsx
